// src/types/store.type.ts
export type StoreType = {
  address: string
  avgReviewStar: number | undefined
  breakStatus: boolean
  description: string
  id: string
  identifier: string[]
  infoBanner: string
  logo: string
  numFollower: number
  numFollowing: number
  numReview: number
  ownerName: string
  shopName: string
  tax: string[]
  topBanner: string
}